Excel Workbook Template
This is a workbook with some model data visualizations in Poverty Solutions style. It serves as a guide and reference last updated in Fall 2022. It is neither an input or an output of any code. [In other words, if the data visualization standards change from the style guide here: https://poverty.umich.edu/files/2017/02/170081_Poverty-Solutions_Look-and-Feel-Guidelines.pdf. 

This file will fall out of date.